cart logic outsourced to useContext
